chore: move rec controller and recorder under one folder
