--- src/recorder/index.ts
@@ -1,2 +1,0 @@
-export * from './recorder';
-export * from './utils';
