+++ src/recording/index.ts
@@ -1,3 +1,4 @@
+export * from './recorder';
 export * from './controller';
 export * from './utils';
 export * from './replace';
