+++ src/recording/utils.ts
@@ -205,3 +205,13 @@ export function checkSensitiveInformation(
     }
   }
 }
+
+export function assertsDefinitionsAreNotStrings(
+  definitions: string[] | RecordingDefinitions
+): asserts definitions is RecordingDefinitions {
+  for (const def of definitions) {
+    if (typeof def === 'string') {
+      throw new UnexpectedError('definition is a string, not object');
+    }
+  }
+}
